refactor(chat): extract chat app into ChatAppMixin (fast_dash/chat_app.py)
Pure move: the ~885 lines of Dash-facing chat application (layout, callbacks,
turn driver, rich-frame rendering, HITL, canvas) move verbatim from fast_dash.py
into a ChatAppMixin that FastDash inherits. fast_dash.py drops 3060 -> 2175
lines; the transport-independent core stays in chat.py. No behavior change;
tests untouched and green (112 chat+mcp).
